scanner/requirements.txt
# Workflow generator + Orchestrator
google-genai>=1.0.0
PyGithub>=2.4.0
httpx>=0.24.0
browser-use-sdk>=3.0.0
uvicorn>=0.22.0
fastapi>=0.100.0
